common: Move memfd ops to shared lib
