+++ credentialsd-ui/Cargo.toml
@@ -16,7 +16,6 @@ futures-lite.workspace = true
 gettext-rs = { version = "0.7", features = ["gettext-system"] }
 gtk = { version = "0.10.3", package = "gtk4", features = ["v4_14"] }
 gdk-wayland = { version = "0.10.3", package = "gdk4-wayland", optional = true }
-libc.workspace = true
 qrcode = "0.14.1"
 serde.workspace = true
 tracing.workspace = true
