+++ credentialsd-ui/src/client.rs
@@ -1,21 +1,13 @@
-use std::{
-    io::ErrorKind,
-    os::{
-        fd::{FromRawFd, OwnedFd},
-        raw::c_void,
-    },
-};
-
 use async_std::{
     channel::{Receiver, Sender},
     sync::Mutex as AsyncMutex,
 };
-use libc::{
-    MAP_SHARED, O_CLOEXEC, PROT_READ, PROT_WRITE, SYS_memfd_secret, ftruncate, mmap, off_t,
-};
-use zeroize::Zeroize;
 
-use credentialsd_common::{model::UserInteractedEvent, server::BackgroundEvent};
+use credentialsd_common::{
+    memfd::write_secret, model::UserInteractedEvent, server::BackgroundEvent,
+};
+
+const CTAP_CLIENT_SECRET_MAX_LEN: usize = 63;
 
 #[derive(Debug)]
 pub struct FlowControlClient {
@@ -29,7 +21,7 @@ impl FlowControlClient {
     }
 
     pub async fn enter_client_pin(&mut self, pin: String) -> Result<(), ()> {
-        let fd = match write_secret(pin) {
+        let fd = match write_secret(pin.into_bytes(), CTAP_CLIENT_SECRET_MAX_LEN) {
             Ok(fd) => fd,
             Err(err) => {
                 tracing::error!(%err, "Failed to write secret to file descriptor");
@@ -65,62 +57,4 @@ impl FlowControlClient {
             Err(_) => Err(()),
         }
     }
-}
-
-fn write_secret(secret: String) -> Result<OwnedFd, std::io::Error> {
-    let mut bytes = secret.into_bytes();
-    // CTAP pins are maximum of 63 bytes, so they should all fit in a u8.
-    let bytes_len = if bytes.len() <= 63 {
-        bytes.len() as u8
-    } else {
-        return Err(std::io::Error::new(
-            ErrorKind::FileTooLarge,
-            "value is too large",
-        ));
-    };
-
-    // Open memfd_secret
-    let ret: i64 = unsafe { libc::syscall(SYS_memfd_secret, O_CLOEXEC) };
-    if ret == -1 {
-        tracing::debug!("Failed to create memfd_secret");
-        return Err(std::io::Error::last_os_error());
-    }
-    let fd = i32::try_from(ret).map_err(|_| std::io::Error::other("invalid file descriptor"))?;
-    if unsafe { ftruncate(fd, bytes_len as off_t) } == -1 {
-        tracing::debug!("Failed to ftruncate memfd_secret");
-        return Err(std::io::Error::last_os_error());
-    }
-
-    let ptr = unsafe {
-        let ptr = mmap(
-            std::ptr::null_mut(),
-            4096,
-            PROT_READ | PROT_WRITE,
-            MAP_SHARED,
-            fd,
-            0,
-        );
-        if ptr == usize::MAX as *mut c_void {
-            tracing::debug!("Failed to mmap memfd_secret");
-            return Err(std::io::Error::last_os_error());
-        }
-        // ptr as *mut u8
-        ptr
-    };
-
-    // Copy the data
-    unsafe {
-        ptr.copy_from_nonoverlapping(bytes.as_ptr().cast(), bytes.len());
-    }
-
-    // Cleanup
-    if unsafe { libc::munmap(ptr, 4096) } == -1 {
-        tracing::debug!("Failed to unmap memfd_secret");
-        return Err(std::io::Error::last_os_error());
-    }
-    bytes.zeroize();
-    drop(bytes);
-
-    let owned_fd = unsafe { OwnedFd::from_raw_fd(fd) };
-    Ok(owned_fd)
 }
